edited code to fix errors on math.h

+++ GameFiles/GameProject/Math.h
@@ -99,7 +99,7 @@ public:
 // calculate distance from 2 points
 float distance(float x1, float y1, float x2, float y2)
 {
-	return sqrt(pow(x2 - x1, 2) + pow(y2 - y1, 2) * 1.0);
+	return static_cast<float>(sqrt(pow(x2 - x1, 2) + pow(y2 - y1, 2) * 1.0));
 }
 
 float clamp(float value, float min, float max)
